refactor: (#3) jpa에서 entity로 패키지 이름 변경
