+++ src/main/java/com/stac/daijin/domain/card/domain/Card.java
@@ -1,7 +1,7 @@
 package com.stac.daijin.domain.card.domain;
 
 import com.stac.daijin.domain.user.domain.User;
-import com.stac.daijin.global.jpa.BaseUUIDEntity;
+import com.stac.daijin.global.entity.BaseUUIDEntity;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
